1 Finishing the edit blog in server side 2 Convert the categoryid's type(byte) to integer because there is chaos in type defination

+++ src/main/java/com/seancheer/dao/entity/BaseCategory.java
@@ -26,7 +26,7 @@ public interface BaseCategory {
 
 	String getDescription();
 
-	Byte getId();
+	Integer getId();
 
 	String getName();
